+++ app/Http/Controllers/ProductController.php
@@ -2,14 +2,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Product;
-use App\Http\Requests\ProfileUpdateRequest;
-use Illuminate\Contracts\Auth\MustVerifyEmail;
-use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Redirect;
 use Inertia\Inertia;
-use Inertia\Response;
 use App\Models\Category;
 use App\Models\ProductType;
 
@@ -17,19 +12,18 @@ class ProductController extends Controller
 {
     public function store(Request $request)
     {
-        $out = new \Symfony\Component\Console\Output\ConsoleOutput();
-        $out->writeln("store");
-        $out->writeln($request->all());
+        if ($request->user()->is_staff != 1) {
+            return Redirect::route("home");
+        }
         $validatedData = $request->validate([
             "name" => "required|string|max:255",
             "description" => "nullable|string",
             "price" => "required|numeric",
+            "discount" => "numeric",
             "category_id" => "required|exists:categories,id",
             "product_type_id" => "required|exists:product_types,id",
             "image" => "nullable|image|mimes:jpg,png,jpeg,gif,svg|max:2048",
         ]);
-
-        $out->writeln($validatedData);
 
         if ($request->hasFile("image")) {
             $imagePath = $request->file("image")->store("products", "public");
@@ -43,12 +37,19 @@ class ProductController extends Controller
             ->with("success", "Product created successfully!");
     }
 
-    public function update(Request $request, Product $product)
+    public function update(Request $request)
     {
+        if ($request->user()->is_staff != 1) {
+            return Redirect::route("home");
+        }
+        
+        $product = Product::find($request->product_id);
+        
         $validatedData = $request->validate([
             "name" => "required|string|max:255",
             "description" => "nullable|string",
             "price" => "required|numeric",
+            "discount" => "numeric",
             "category_id" => "required|exists:categories,id",
             "product_type_id" => "required|exists:product_types,id",
             "image" => "nullable|image|mimes:jpg,png,jpeg,gif,svg|max:2048",
@@ -62,12 +63,15 @@ class ProductController extends Controller
         $product->update($validatedData);
 
         return redirect()
-            ->route("products")
+            ->route("products.view", ["id" => $product->id])
             ->with("success", "Product updated successfully!");
     }
 
-    public function destroy(Product $product)
+    public function destroy(Request $request, Product $product)
     {
+        if ($request->user()->is_staff != 1) {
+            return Redirect::route("home");
+        }
         $product->delete();
 
         return redirect()
@@ -75,8 +79,11 @@ class ProductController extends Controller
             ->with("success", "Product deleted successfully!");
     }
 
-    public function list()
+    public function list(Request $request)
     {
+        if ($request->user()->is_staff != 1) {
+            return Redirect::route("home");
+        }
         $products = Product::with(["category", "product_type"])->get();
         $categories = Category::all();
         $productTypes = ProductType::all();
@@ -90,9 +97,46 @@ class ProductController extends Controller
 
     public function view($id)
     {
-        $product = Product::find($id);
+        $product = Product::with(["category", "product_type"])->find($id);
         return Inertia::render("Product/View", [
             "product" => $product,
+        ]);
+    }
+
+    public function filter(Request $request)
+    {
+        $out = new \Symfony\Component\Console\Output\ConsoleOutput();
+        // $out->writeln("store");
+        // $out->writeln($request->all());
+
+        // Retrieve the selected category and product type IDs from the request
+        $categoryIds = $request->input("category", []); // Defaults to an empty array if not provided
+        $productTypeIds = $request->input("product_type", []); // Defaults to an empty array if not provided
+
+        // Check if both arrays are empty, and return an empty products array if true
+        if (empty($categoryIds) && empty($productTypeIds)) {
+            $products = [];
+        } else {
+            // Filter products based on the given category and product type IDs
+            $products = Product::with(["category", "product_type"])
+                ->when(!empty($categoryIds), function ($query) use (
+                    $categoryIds
+                ) {
+                    return $query->whereIn("category_id", $categoryIds);
+                })
+                ->when(!empty($productTypeIds), function ($query) use (
+                    $productTypeIds
+                ) {
+                    return $query->whereIn("product_type_id", $productTypeIds);
+                })
+                ->get();
+        }
+
+        // Return an Inertia response with the filtered products
+        return Inertia::render("Home", [
+            "products" => $products,
+            "categories" => Category::all(),
+            "product_types" => ProductType::all(),
         ]);
     }
 }
